+++ dependencies/imgui/addons/imguivariouscontrols/imguivariouscontrols.h
@@ -430,7 +430,7 @@ public:
             const char e = '\0';
             if (!text && !allowNullDestText) text=&e;
             if (text)  {
-                const int sz = strlen(text);
+                const int sz = (int)strlen(text);
                 destText = (char*) ImGui::MemAlloc(sz+1);strcpy(destText,text);
             }
         }
